+++ src/lib.rs
@@ -35,6 +35,6 @@ mod parse;
 mod screen;
 mod style;
 
-pub use parse::{Event, Parser, Token};
+pub use parse::{Event, Parser, Token, Utf8Decoder};
 pub use screen::{Cell, CellWidth, Cursor, Screen};
 pub use style::{Color, Style};